third_party/WebKit/LayoutTests/external/wpt/dom/events/EventListener-incumbent-global-2.sub-expected.txt

This is a testharness.js-based test.
FAIL Check the incumbent global EventListeners  are called withassert_equals: Incumbent should have been the caller of addEventListener() expected "http://www1.web-platform.test:8001" but got "http://www2.web-platform.test:8001"
Harness: the test ran to completion.
